requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
argon2-cffi==23.1.0
requests==2.31.0
firebase-admin==6.5.0
python-dotenv==1.0.0

